requirements.txt
aiohttp==3.8.4
asyncio==3.4.3
certifi>=2022.12.7
google-cloud-bigquery>=3.12.0,<4.0.0
google-cloud-storage>=2.11.0,<3.0.0
google-auth>=2.13.0,<3.0.0
httplib2==0.22.0
matplotlib==3.8.0
numpy>=1.26.0,<2.0.0
pandas>=2.1.2,<3.0.0
pandas-gbq>=0.19.0,<1.0.0
Pillow==10.1.0
python-dotenv>=1.0.0,<2.0.0
qdrant-client>=1.5.0,<2.0.0
requests>=2.31.0,<3.0.0
sentence-transformers>=3.0.1,<4.0.0
huggingface_hub>=0.12.1,<1.0.0
tabulate==0.9.0
tenacity>=8.2.3,<9.0.0
tqdm>=4.66.1,<5.0.0
pydantic>=1.10.9,<2.0.0
pydrive2>=1.14.0,<2.0.0
gspread>=5.8.0,<6.0.0
torch>=2.0.0,<3.0.0
